twsdo: remove OPT_PARAMS_END debug message
... and minor cleanup.

+++ src/twsdo.cpp
@@ -479,8 +479,7 @@ bool TwsDL::finContracts()
 	PacketContractDetails* p = (PacketContractDetails*)packet;
 	switch( packet->getError() ) {
 	case REQ_ERR_NONE:
-		DEBUG_PRINTF( "Contracts received: %lu",
-			(unsigned long) p->constList().size() );
+		DEBUG_PRINTF("Contracts received: %zu", p->constList().size());
 		packet->dumpXml();
 	case REQ_ERR_NODATA:
 	case REQ_ERR_NAV:
@@ -515,8 +514,7 @@ bool TwsDL::finOptParams()
 	PacketOptParams* p = (PacketOptParams*)packet;
 	switch( packet->getError() ) {
 	case REQ_ERR_NONE:
-		DEBUG_PRINTF( "OptParams received: %lu",
-			(unsigned long) p->constList().size() );
+		DEBUG_PRINTF("OptParams received: %zu", p->constList().size());
 		packet->dumpXml();
 	case REQ_ERR_NODATA:
 	case REQ_ERR_NAV:
